updates to rpc.sendOperation function
In all places where you set the variable of operation in order to utilize it in the rpc.sendOperation method you need to turn that object into an array with an object inside. Otherwise, operation won't be iterable when you use the spread operation for the variable ops in the rpc.sendOperation function.